pyproject.toml
[project]
name = "intellimail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.12",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.11",
    "python-dotenv>=1.2.2",
]


